+++ packages/embed-wallet/src/types.ts
@@ -41,6 +41,7 @@ export type UserInfo = {
   name: string;
   profileImage: string;
   isNewUser: boolean;
+  isNewWallet: boolean;
 };
 
 export type WalletAccountType = 'ethereum' | 'ed25519' | 'solana';
@@ -73,6 +74,7 @@ export type WalletConnectOptions = {
   redirectUrl?: string;
   permissions?: PermissionRequest;
   loginHint?: string;
+  email?: string;
 };
 
 export type TokenType = 'erc20' | 'native';
@@ -139,13 +141,18 @@ export type PermissionCaveat = {
   value: any;
 };
 
+type KnownPermissionCaveats = {
+  title?: string;
+  description?: string | null;
+};
+
 export type Permission = {
   parentCapability: string;
   caveats: PermissionCaveat[];
 };
 
 export type PermissionRequest = {
-  [methodName: Permission['parentCapability']]: {
+  [methodName: Permission['parentCapability']]: KnownPermissionCaveats & {
     [caveatName: PermissionCaveat['type']]: PermissionCaveat['value'];
   };
 };