+++ src/components/Permissions/Permissions.tsx
@@ -1,7 +1,7 @@
 import { List, ListItem, ListItemText, Checkbox, ListItemIcon } from '@cere-wallet/ui';
 import type { PermissionRequest } from '@cere-wallet/wallet-engine';
 
-import { AllowedPermission, defaultDescription, knownPermissions } from './knownPermissions';
+import { AllowedPermission, knownPermissions } from './knownPermissions';
 
 export type PermissionsProps = {
   permissions: PermissionRequest;
@@ -18,20 +18,22 @@ export const Permissions = ({ value, permissions, onChange }: PermissionsProps) 
 
   return (
     <List dense disablePadding>
-      {Object.keys(permissions).map((capability) => (
-        <ListItem key={capability} disableGutters disablePadding>
-          <ListItemIcon>
-            <Checkbox
-              checked={!!value[capability]}
-              onChange={(event) => updateSelection(capability, event.target.checked)}
-            />
-          </ListItemIcon>
-          <ListItemText
-            primary={knownPermissions[capability as AllowedPermission]?.title || capability}
-            secondary={knownPermissions[capability as AllowedPermission]?.description || defaultDescription}
-          />
-        </ListItem>
-      ))}
+      {Object.keys(permissions).map((capability) => {
+        const defaults = knownPermissions[capability as AllowedPermission] || {};
+        const permission = { ...defaults, ...permissions[capability as AllowedPermission] };
+
+        return (
+          <ListItem key={capability} disableGutters disablePadding>
+            <ListItemIcon>
+              <Checkbox
+                checked={!!value[capability]}
+                onChange={(event) => updateSelection(capability, event.target.checked)}
+              />
+            </ListItemIcon>
+            <ListItemText primary={permission.title || capability} secondary={permission.description} />
+          </ListItem>
+        );
+      })}
     </List>
   );
 };