+++ src/index.tsx
@@ -1,13 +1,8 @@
 import ReactDOM from 'react-dom/client';
-import TagManager from 'react-gtm-module';
 
 import Reporting from './reporting';
-import { GTM_ID, APP_VERSION } from './constants';
+import { APP_VERSION } from './constants';
 import App from './App';
-
-if (GTM_ID) {
-  TagManager.initialize({ gtmId: GTM_ID });
-}
 
 /**
  * Log the version of the wallet client