+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 -
 
+### v1.39.0
+
+- Add support for externally provided email to login with
+- Make it possible to change permission title and description
+- Store wallet permission to connected application and don't ask on each login
+- Add `Magic Link` authentication support
+- Postpone GTM initialization to increase init speed
+- Add transaction signing permission
+
 ### v1.38.0
 
 - Added multi-chain signature API to the wallet client and SDK